rename l-r-d to mwc-l-r-d
