+++ packages/textfield/src/mwc-textfield-base.ts
@@ -18,7 +18,7 @@ import '@material/mwc-notched-outline';
 
 import {addHasRemoveClass, classMap, FormElement, html, property, query, TemplateResult} from '@material/mwc-base/form-element.js';
 import {floatingLabel, FloatingLabel} from '@material/mwc-floating-label/mwc-floating-label-directive';
-import {lineRipple, LineRipple} from '@material/mwc-line-ripple/line-ripple-directive.js';
+import {lineRipple, LineRipple} from '@material/mwc-line-ripple/mwc-line-ripple-directive.js';
 import {NotchedOutline} from '@material/mwc-notched-outline';
 import {MDCTextFieldAdapter, MDCTextFieldInputAdapter, MDCTextFieldLabelAdapter, MDCTextFieldLineRippleAdapter, MDCTextFieldOutlineAdapter, MDCTextFieldRootAdapter} from '@material/textfield/adapter.js';
 import MDCTextFieldFoundation from '@material/textfield/foundation.js';
